Move function from args to alongside caller in main
